kvstreamer: account for the overhead of each request
Previously, we didn't account for some of the overhead of Get and Scan
requests, and this commit fixes that.

Here is the list of all things that we need to account for a single Get
request:
- (if referenced by `[]roachpb.RequestUnion`) the overhead of the
`roachpb.RequestUnion` object
- the overhead of `roachpb.RequestUnion_Get` object
- the overhead of `roachpb.GetRequest` object
- the footprint of the key inside of `roachpb.GetRequest`.

Previously, we only accounted for the first and the fourth items, and
now we account for all of them. Additionally, we used the
auto-generated `Size` method for the fourth item which I believe
represent the size of the serialized protobuf message (possibly
compressed - I'm not sure), but we're interested in the capacities of
the underlying slices.

Release note: None
