+++ pkg/kv/kvclient/kvstreamer/BUILD.bazel
@@ -7,6 +7,7 @@ go_library(
         "budget.go",
         "requests_provider.go",
         "results_buffer.go",
+        "size.go",
         "streamer.go",
     ],
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvclient/kvstreamer",
